Set timeout and Button disable for proper completion of game.
I got a tip that I should disable all the buttons until it player doesn't reset to play a new game.
I did this using .disabled method.
Inorder to display who won the last turn,I set a timeout of half-second to display it in paragraph
and then let alert come out after a while to show who won the overall round.
This helps in proper completion of the game.